refactor of the OnboardingService to unify web and app
