+++ src/main/java/it/gov/pagopa/onboarding/workflow/controller/OnboardingControllerImpl.java
@@ -1,6 +1,9 @@
 package it.gov.pagopa.onboarding.workflow.controller;
 
 import it.gov.pagopa.onboarding.workflow.dto.*;
+import it.gov.pagopa.onboarding.workflow.dto.web.InitiativeWebDTO;
+import it.gov.pagopa.onboarding.workflow.enums.ChannelType;
+import it.gov.pagopa.onboarding.workflow.exception.custom.InitiativeNotFoundException;
 import it.gov.pagopa.onboarding.workflow.service.OnboardingService;
 import jakarta.validation.Valid;
 import org.springframework.data.domain.Pageable;
@@ -10,6 +13,9 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.time.LocalDateTime;
+import java.util.Locale;
+
+import static it.gov.pagopa.onboarding.workflow.constants.OnboardingWorkflowConstants.ExceptionMessage.INITIATIVE_NOT_FOUND_MSG;
 
 @RestController
 public class OnboardingControllerImpl implements OnboardingController {
@@ -32,9 +38,9 @@ public class OnboardingControllerImpl implements OnboardingController {
   }
 
   public ResponseEntity<Void> onboardingCitizen(
-      @Valid @RequestBody OnboardingPutDTO onBoardingPutDTO,
+      @Valid @RequestBody OnboardingSaveDTO onBoardingSaveDTO,
       String userId) {
-    onboardingService.putTcConsent(onBoardingPutDTO.getInitiativeId(),userId);
+    onboardingService.putTcConsent(onBoardingSaveDTO.getInitiativeId(),userId);
     return new ResponseEntity<>(HttpStatus.NO_CONTENT);
   }
 
@@ -87,6 +93,24 @@ public class OnboardingControllerImpl implements OnboardingController {
   public  ResponseEntity<OnboardingFamilyDTO> familyUnitComposition(String initiativeId, String userId){
     OnboardingFamilyDTO onboardingFamilyDTO = onboardingService.getfamilyUnitComposition(initiativeId, userId);
     return new ResponseEntity<>(onboardingFamilyDTO, HttpStatus.OK);
+  }
+
+  @Override
+  public ResponseEntity<InitiativeWebDTO> getInitiative(
+          String initiativeId, Locale acceptLanguage) {
+    InitiativeWebDTO dto = onboardingService.initiativeDetail(initiativeId, acceptLanguage);
+    if (dto == null) {
+      throw new InitiativeNotFoundException(String.format(INITIATIVE_NOT_FOUND_MSG, initiativeId), true, null);
+
+    }
+    return ResponseEntity.ok(dto);
+  }
+
+  @Override
+  public ResponseEntity<Void> saveConsent(ConsentPutDTO consentPutDTO, ChannelType channel, String userId) {
+    consentPutDTO.setChannel(channel);
+    onboardingService.saveOnboarding(consentPutDTO, userId);
+    return ResponseEntity.accepted().build();
   }
   
 }
