+++ src/main/java/it/gov/pagopa/onboarding/workflow/dto/ConsentPutDTO.java
@@ -1,6 +1,8 @@
 package it.gov.pagopa.onboarding.workflow.dto;
 
+import it.gov.pagopa.onboarding.workflow.enums.ChannelType;
 import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -9,16 +11,31 @@ import lombok.NoArgsConstructor;
 import java.util.List;
 
 @Data
-@Builder
 @NoArgsConstructor
 @AllArgsConstructor
+@Builder
 public class ConsentPutDTO {
 
-  @NotBlank(message="The field is mandatory")
-  String initiativeId;
+    @NotBlank(message = "The initiativeId is mandatory")
+    private String initiativeId;
 
-  boolean pdndAccept;
+    private boolean pdndAccept;
 
-  List<SelfConsentDTO> selfDeclarationList;
+    private List<SelfConsentDTO> selfDeclarationList;
+
+
+    private String userMail;
+    private String userMailConfirmation;
+    private Boolean confirmedTos;
+
+    @NotNull(message = "Channel is mandatory")
+    private ChannelType channel;
+
+    public boolean isWebChannel() {
+        return ChannelType.WEB == channel;
+    }
+
+    public boolean isIoChannel() {
+        return ChannelType.APP_IO == channel;
+    }
 }
-
